+++ src/lightly_train/_commands/train_task_helpers.py
@@ -14,7 +14,7 @@ import logging
 from functools import partial
 from json import JSONEncoder
 from pathlib import Path
-from typing import Any, Generator, Iterable, Literal
+from typing import Any, Generator, Iterable, Literal, Mapping
 
 import torch
 from filelock import FileLock
@@ -26,8 +26,11 @@ from torch.utils.data import DataLoader, Dataset
 
 from lightly_train._configs import validate
 from lightly_train._data import cache
-from lightly_train._data._serialize import memory_mapped_sequence
-from lightly_train._data._serialize.memory_mapped_sequence import MemoryMappedSequence
+from lightly_train._data._serialize import memory_mapped_sequence_task
+from lightly_train._data._serialize.memory_mapped_sequence_task import (
+    MemoryMappedSequenceTask,
+    Primitive,
+)
 from lightly_train._data.mask_semantic_segmentation_dataset import (
     MaskSemanticSegmentationDataset,
     MaskSemanticSegmentationDatasetArgs,
@@ -358,12 +361,12 @@ def _decrement_and_cleanup_if_zero(mmap_file: Path, ref_file: Path) -> None:
         pass  # Another process already cleaned up
 
 
-def get_dataset_mmap_filenames(
+def get_dataset_mmap_file(
     fabric: Fabric,
-    filenames: Iterable[str],
+    items: Iterable[Mapping[str, Primitive]],
     mmap_filepath: Path,
-) -> MemoryMappedSequence[str]:
-    """Returns memory-mapped filenames shared across all ranks.
+) -> MemoryMappedSequenceTask[Primitive]:
+    """Returns memory-mapped filepaths shared across all ranks.
 
     Filenames are written to mmap_filepath by rank zero and read by all ranks.
     """
@@ -371,9 +374,7 @@ def get_dataset_mmap_filenames(
     # If the file already exists and we are allowed to reuse it, return it.
     if Env.LIGHTLY_TRAIN_MMAP_REUSE_FILE.value and mmap_filepath.exists():
         logger.warning(f"Reusing existing memory-mapped file '{mmap_filepath}'.")
-        return memory_mapped_sequence.memory_mapped_sequence_from_file(
-            mmap_filepath=mmap_filepath
-        )
+        return MemoryMappedSequenceTask.from_file(mmap_filepath=mmap_filepath)
 
     # Check if the mmap file is on a shared filesystem.
     try:
@@ -390,15 +391,12 @@ def get_dataset_mmap_filenames(
         if (fabric.global_rank == 0) or (
             not is_shared_filesystem and fabric.local_rank == 0
         ):
-            memory_mapped_sequence.write_filenames_to_file(
-                filenames=filenames,
+            memory_mapped_sequence_task.write_items_to_file(
+                items=items,
                 mmap_filepath=mmap_filepath,
             )
 
-    # Return memory-mapped filenames from file.
-    return memory_mapped_sequence.memory_mapped_sequence_from_file(
-        mmap_filepath=mmap_filepath
-    )
+    return MemoryMappedSequenceTask.from_file(mmap_filepath=mmap_filepath)
 
 
 def get_dataset(
@@ -407,13 +405,14 @@ def get_dataset(
     transform: TaskTransform,
     mmap_filepath: Path,
 ) -> MaskSemanticSegmentationDataset:
-    filenames = list(dataset_args.list_image_filenames())
+    image_info = dataset_args.list_image_info()
+
     dataset_cls = dataset_args.get_dataset_cls()
     return dataset_cls(
         dataset_args=dataset_args,
-        image_filenames=get_dataset_mmap_filenames(
+        image_info=get_dataset_mmap_file(
             fabric=fabric,
-            filenames=filenames,
+            items=image_info,
             mmap_filepath=mmap_filepath,
         ),
         transform=transform,